Add pausing between painting each box
Value Add:
It will be easier for students to grasp that graphics are painted one at
a time, in the order they specify.
This becomes important when drawing overlapping shapes.

Use Cases:
The assignment is to draw three overlapping boxes.
If you position them as the assignment specifies but draw them in the
wrong order, one or more boxes will not be visible.
This often confuses students.

Rationale for Choices:
The waiting code is inside a method so as not to distract or confuse
students.
Drawing additional text after every rectangle that says, "waiting before
drawing next rectangle" requires additional logic to draw this text at
a different location each time or drawing a rectangle (but of what
color?) behind the text each time. This is either (1) tricky to write so
as to cover all student use cases, or (2) messy on the canvas, causing
distraction or frustration.